requirements.txt
Django==1.11.7
Pillow==2.9.0
arrow==	0.10.0
beautifulsoup4==4.6.0
certifi==2017.11.5
chardet==3.0.4
django-bootstrap==0.2.4
django-carton==1.2.1
django-debug-toolbar==1.9.1
django-disqus==0.5
django-emoticons==1.2
django-forms-bootstrap==3.1.0
django-paypal==0.3.6
django-tinymce==2.6.0
funcsigs==1.0.2
idna==2.6
mock==2.0.0
pbr==3.1.1
pip==9.0.1
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
setuptools==36.8.0
six==1.11.0
sqlparse==0.2.4
stripe==1.75.0
urllib3==1.22
wheel==0.30.0
gunicorn==19.4
